test(app): make TestE2E_PaneFlow robust to a reordering focus ring (#2559)
TestE2E_PaneFlow timed out on the required Linux gate (pane_model_test.go:1449,
"x hides the focused pane") on a docs-only PR — byte-identical code, so a
behavioral regression is impossible. Reproduced deterministically under a starved
scheduler (GOMAXPROCS=1) and, crucially, on the ANCESTOR before the four
pane/TUI commits merged in the prior 18h — the same signature — so it predates
them and is not a regression (the #2052 timing-flake class, now on Linux; #1611
hardened the same flows for darwin).

Mechanism (measured, not argued): the focus ring's pane entries are recency-ranked
(store.VisibleOpenPanes), so any pane touch — including the 100ms preview-tick
relayout — REORDERS the ring. The test drove a blind, fixed-count 4-Tab burst then
asserted the ring was on a pane; under contention a background relayout reorders the
ring between the keys and the assertion, so the 4th Tab lands on `projects` (a
non-pane) and sits there for the full 30s, and the later `x` then hides nothing. A
per-Tab trace was decisive: one Tab from `pane:2` jumped straight to `tree`,
proving the ring order shifted under it.

Fix (test-only, no product change, no wider timeout, no sleeps-that-wait-longer):
- Assert Tab MOVES the ring off its current stop rather than asserting a fixed
  traversal count — the count is unknowable when the ring reorders under load.
- Hide with retry-until-effect: re-drive the ring onto a pane and re-send `x`
  until a pane is actually hidden, since a relayout can move the ring off the
  focused pane between the key and its Update, making a single `x` a no-op. Bounded
  by the same 30s ceiling; a working `x` breaks out at once, only a no-op re-checks.

Proof: survives `make test-container GOTESTARGS="-run TestE2E_PaneFlow -count=2000
-cpu 1"` (2000/2000), where the unfixed test fails ~11/2000 on master and ~1/1000
on the ancestor.

Scope: app E2E family only. The mitigation gap this exposes (see #2559: `-short`
covers build.yml but not the required pr.yml gates, and only ever covered app E2E)
is tracked there, including a separate apiclient socket-dial instance
(TestPreviewSnapshotCarriesTerminalModes) that is the same family but a different
mechanism and needs a different fix.
