+++ app/pane_model_test.go
@@ -1377,81 +1377,6 @@ func TestPane_HideMiddleFocusesSuccessor(t *testing.T) {
 		"focus lands on the pane that took the hidden pane's slot")
 }
 
-// TestE2E_PaneFlow drives the real tea.Program through the pane lifecycle:
-// s opens a focused pane, the tree walks to another instance and s opens a
-// second pane beside it, Tab cycles the ring across both, and x hides the
-// focused pane with focus landing on the survivor.
-func TestE2E_PaneFlow(t *testing.T) {
-	eh := newE2EHarness(t)
-	eh.addStartedInstance("alpha")
-	eh.addStartedInstance("beta")
-	eh.home.sidebar.SetSelectedInstance(0)
-	eh.start()
-
-	paneState := func() (count int, titles []string, region string) {
-		eh.query(func(h *home) {
-			count = h.store.NumOpenPanes()
-			titles = visibleTitles(h)
-			region = h.ring.Active()
-		})
-		return
-	}
-
-	// s opens the selection (alpha) as a focused pane.
-	eh.tm.Send(tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune{'s'}})
-	eh.waitUntil(e2eAsyncTimeout, "s opens alpha's pane focused", func() bool {
-		count, titles, region := paneState()
-		return count == 1 && len(titles) == 1 && titles[0] == "alpha" && layout.IsPaneRegion(region)
-	})
-
-	// The tree walks to beta (j walks alpha → its two tab rows → beta);
-	// alpha's pane stays put. Then s opens beta beside it.
-	for i := 0; i < 3; i++ {
-		eh.tm.Send(tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune{'j'}})
-	}
-	eh.waitUntil(e2eAsyncTimeout, "tree selection lands on beta", func() bool {
-		var selected string
-		eh.query(func(h *home) {
-			if s := h.store.GetSelectedInstance(); s != nil {
-				selected = s.Title
-			}
-		})
-		return selected == "beta"
-	})
-	eh.tm.Send(tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune{'s'}})
-	eh.waitUntil(e2eAsyncTimeout, "s opens beta's pane to the right", func() bool {
-		count, titles, _ := paneState()
-		return count == 2 && len(titles) == 2 && titles[0] == "alpha" && titles[1] == "beta"
-	})
-
-	// Both pane headers render side by side.
-	var view string
-	eh.query(func(h *home) { view = h.View() })
-	// Pane 2's tab label is whatever the shared active-tab index resolved to
-	// after the tree walk (the index survives instance switches by design),
-	// so assert the instance halves only.
-	assert.Contains(t, view, "alpha · ◆ Agent", "pane 1 header shows its binding")
-	assert.Contains(t, view, "beta · ", "pane 2 header shows its binding")
-
-	// Tab from the beta pane wraps via automations/projects/tree back around to
-	// the alpha pane; assert the ring visits a pane region again.
-	eh.tm.Send(tea.KeyMsg{Type: tea.KeyTab})
-	eh.tm.Send(tea.KeyMsg{Type: tea.KeyTab})
-	eh.tm.Send(tea.KeyMsg{Type: tea.KeyTab})
-	eh.tm.Send(tea.KeyMsg{Type: tea.KeyTab})
-	eh.waitUntil(e2eAsyncTimeout, "the focus ring cycles back to a pane", func() bool {
-		_, _, region := paneState()
-		return layout.IsPaneRegion(region)
-	})
-
-	// x hides the focused pane; the survivor keeps focus and the workspace.
-	eh.tm.Send(tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune{'x'}})
-	eh.waitUntil(e2eAsyncTimeout, "x hides the focused pane", func() bool {
-		count, titles, region := paneState()
-		return count == 1 && len(titles) == 1 && layout.IsPaneRegion(region)
-	})
-}
-
 // TestPane_StoreOpenPanePrimitives unit-tests the store's open-pane list:
 // dedupe lookup, ordered append, close, and the recency-ranked visibility
 // pick that drives the §2.6 auto-hide.
